Renamed contracts/contractErrors.ts to contracts/errors.ts
